+++ src/main/java/vazkii/patchouli/client/base/ClientInitializer.java
@@ -3,18 +3,17 @@ package vazkii.patchouli.client.base;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.renderer.item.ItemProperties;
 import net.minecraft.client.renderer.item.ItemPropertyFunction;
+import net.minecraft.client.resources.model.BakedModel;
 import net.minecraft.client.resources.model.ModelResourceLocation;
 import net.minecraft.resources.ResourceLocation;
-import net.minecraft.server.packs.resources.ReloadableResourceManager;
-import net.minecraft.server.packs.resources.ResourceManager;
 import net.minecraft.server.packs.resources.ResourceManagerReloadListener;
-import net.minecraftforge.client.event.ClientPlayerNetworkEvent;
+
+import net.minecraftforge.client.event.ModelBakeEvent;
 import net.minecraftforge.client.event.ModelRegistryEvent;
-import net.minecraftforge.client.event.RenderTooltipEvent;
-import net.minecraftforge.client.event.RenderWorldLastEvent;
+import net.minecraftforge.client.event.RegisterClientReloadListenersEvent;
 import net.minecraftforge.client.model.ModelLoader;
-import net.minecraftforge.common.MinecraftForge;
-import net.minecraftforge.event.TickEvent;
+import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
+import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 
 import vazkii.patchouli.client.book.ClientBookRegistry;
 import vazkii.patchouli.client.handler.BookRightClickHandler;
@@ -24,57 +23,50 @@ import vazkii.patchouli.common.base.Patchouli;
 import vazkii.patchouli.common.book.BookRegistry;
 import vazkii.patchouli.common.item.ItemModBook;
 import vazkii.patchouli.common.item.PatchouliItems;
-import vazkii.patchouli.common.network.NetworkHandler;
 
-public class ClientProxy {
-	public void onInitializeClient() {
+public class ClientInitializer {
+	public static void onInitializeClient(FMLClientSetupEvent evt) {
 		ClientBookRegistry.INSTANCE.init();
 		PersistentData.setup();
-		ClientTicker.init();
 		BookRightClickHandler.init();
 		MultiblockVisualizationHandler.init();
-		NetworkHandler.registerMessages();
+		TooltipHandler.init();
+		ClientAdvancements.init();
+		ClientTicker.init();
 
-		MinecraftForge.EVENT_BUS.addListener((ModelRegistryEvent e) -> {
-			BookRegistry.INSTANCE.books.values().stream()
-					.map(b -> new ModelResourceLocation(b.model, "inventory"))
-					.forEach(ModelLoader::addSpecialModel);
+		var modEventBus = FMLJavaModLoadingContext.get().getModEventBus();
+		modEventBus.addListener(ClientInitializer::onRegisterClientReloadListener);
+		modEventBus.addListener(ClientInitializer::onModelBake);
+		modEventBus.addListener(ClientInitializer::onModelRegister);
 
-			ItemPropertyFunction prop = (stack, world, entity, seed) -> ItemModBook.getCompletion(stack);
-			ItemProperties.register(PatchouliItems.book, new ResourceLocation(Patchouli.MOD_ID, "completion"), prop);
-		});
+		evt.enqueueWork(() -> ClientBookRegistry.INSTANCE.reload(false));
+	}
 
-		var resourceManager = ((ReloadableResourceManager) Minecraft.getInstance().getResourceManager());
-		resourceManager.registerReloadListener(new ResourceManagerReloadListener() {
-			@Override
-			public void onResourceManagerReload(ResourceManager manager) {
-				if (Minecraft.getInstance().level != null) {
-					Patchouli.LOGGER.info("Reloading resource pack-based books, world is nonnull");
-					ClientBookRegistry.INSTANCE.reload(true);
-				} else {
-					Patchouli.LOGGER.info("Not reloading resource pack-based books as client world is missing");
-				}
-			}
-		});
-
-		MinecraftForge.EVENT_BUS.addListener((TickEvent.RenderTickEvent e) -> {
-			if (e.phase == TickEvent.Phase.START) {
-				ClientTicker.renderTickStart(e.renderTickTime);
+	private static void onRegisterClientReloadListener(RegisterClientReloadListenersEvent evt) {
+		evt.registerReloadListener((ResourceManagerReloadListener) (manager) -> {
+			if (Minecraft.getInstance().level != null) {
+				Patchouli.LOGGER.info("Reloading resource pack-based books, world is nonnull");
+				ClientBookRegistry.INSTANCE.reload(true);
 			} else {
-				ClientTicker.renderTickEnd();
+				Patchouli.LOGGER.info("Not reloading resource pack-based books as client world is missing");
 			}
 		});
+	}
 
-		MinecraftForge.EVENT_BUS.addListener((RenderWorldLastEvent e) -> {
-			MultiblockVisualizationHandler.onWorldRenderLast(e.getMatrixStack());
-		});
+	private static void onModelBake(ModelBakeEvent evt) {
+		ModelResourceLocation key = new ModelResourceLocation(PatchouliItems.BOOK_ID, "inventory");
+		BakedModel oldModel = evt.getModelRegistry().get(key);
+		if (oldModel != null) {
+			evt.getModelRegistry().put(key, new BookModel(oldModel, evt.getModelLoader()));
+		}
+	}
 
-		MinecraftForge.EVENT_BUS.addListener((ClientPlayerNetworkEvent.LoggedOutEvent e) -> {
-			ClientAdvancements.playerLogout();
-		});
+	private static void onModelRegister(ModelRegistryEvent e) {
+		BookRegistry.INSTANCE.books.values().stream()
+				.map(b -> new ModelResourceLocation(b.model, "inventory"))
+				.forEach(ModelLoader::addSpecialModel);
 
-		MinecraftForge.EVENT_BUS.addListener((RenderTooltipEvent.Pre e) -> {
-			TooltipHandler.onTooltip(e.getMatrixStack(), e.getStack(), e.getX(), e.getY());
-		});
+		ItemPropertyFunction prop = (stack, world, entity, seed) -> ItemModBook.getCompletion(stack);
+		ItemProperties.register(PatchouliItems.BOOK, new ResourceLocation(Patchouli.MOD_ID, "completion"), prop);
 	}
 }
